+++ src/main/java/com/mraof/minestuck/computer/ButtonListProgram.java
@@ -2,10 +2,10 @@ package com.mraof.minestuck.computer;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
+import com.mraof.minestuck.blockentity.ComputerBlockEntity;
 import com.mraof.minestuck.client.gui.ComputerScreen;
 import com.mraof.minestuck.network.MSPacketHandler;
 import com.mraof.minestuck.network.computer.ClearMessagePacket;
-import com.mraof.minestuck.blockentity.ComputerBlockEntity;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.components.Button;
@@ -47,11 +47,8 @@ public abstract class ButtonListProgram extends ComputerProgram
 	public final void onButtonPressed(ComputerScreen screen, Button button)
 	{
 		UnlocalizedString data = buttonMap.get(button);
-		if(button == upButton)
-			index--;
-		else if(button == downButton)
-			index++;
-		else if(data != null)
+		
+		if(data != null)
 		{
 			if(!screen.be.latestmessage.get(this.getId()).isEmpty())
 				MSPacketHandler.sendToServer(new ClearMessagePacket(screen.be.getBlockPos(), this.getId()));
@@ -60,21 +57,25 @@ public abstract class ButtonListProgram extends ComputerProgram
 		screen.updateGui();
 	}
 	
+	public final void onArrowPressed(ComputerScreen screen, boolean reverse)
+	{
+		if(reverse) index--; else index++;
+		
+		screen.updateGui();
+	}
+	
 	@Override
 	public final void onInitGui(ComputerScreen gui)
 	{
+		var xOffset = (gui.width-ComputerScreen.xSize)/2;
+		var yOffset = (gui.height-ComputerScreen.ySize)/2;
 		buttonMap.clear();
 		for(int i = 0; i < 4; i++)
-		{
-			Button button = new ExtendedButton((gui.width - ComputerScreen.xSize) / 2 + 14, (gui.height - ComputerScreen.ySize) / 2 + 60 + i * 24, 120, 20, Component.empty(), button1 -> onButtonPressed(gui, button1));
-			buttonMap.put(button, new UnlocalizedString(""));
-			gui.addRenderableWidget(button);
-		}
+			buttonMap.put(gui.addRenderableWidget(
+					new ExtendedButton(xOffset + 14, yOffset + 60 + i * 24, 120, 20, Component.empty(), button -> onButtonPressed(gui, button))), new UnlocalizedString(""));
 		
-		upButton = new ExtendedButton((gui.width - ComputerScreen.xSize) / 2 + 140, (gui.height - ComputerScreen.ySize) / 2 + 60, 20, 20, Component.literal("^"), button1 -> onButtonPressed(gui, button1));
-		gui.addRenderableWidget(upButton);
-		downButton = new ExtendedButton((gui.width - ComputerScreen.xSize) / 2 + 140, (gui.height - ComputerScreen.ySize) / 2 + 132, 20, 20, Component.literal("v"), button1 -> onButtonPressed(gui, button1));
-		gui.addRenderableWidget(downButton);
+		upButton = gui.addRenderableWidget(new ArrowButton(true, gui));
+		downButton = gui.addRenderableWidget(new ArrowButton(false, gui));
 	}
 	
 	@Override
@@ -85,6 +86,7 @@ public abstract class ButtonListProgram extends ComputerProgram
 		ArrayList<UnlocalizedString> list = getStringList(gui.be);
 		if(!gui.be.latestmessage.get(this.getId()).isEmpty())
 			list.add(1, new UnlocalizedString(CLEAR_BUTTON));
+		
 		int pos = -1;
 		for(UnlocalizedString s : list)
 		{
@@ -107,6 +109,7 @@ public abstract class ButtonListProgram extends ComputerProgram
 			}
 			pos++;
 		}
+		
 		if(index == 0 && pos != 4)
 			for(; pos < 4; pos++)
 			{
@@ -125,24 +128,17 @@ public abstract class ButtonListProgram extends ComputerProgram
 	@Override
 	public final void paintGui(PoseStack poseStack, ComputerScreen gui, ComputerBlockEntity be)
 	{
-		int yOffset = (gui.height / 2) - (ComputerScreen.ySize / 2);
-		
-		RenderSystem.setShader(GameRenderer::getPositionTexShader);
-		RenderSystem.setShaderColor(1, 1, 1, 1);
-		RenderSystem.setShaderTexture(0, ComputerScreen.guiBackground);
-		gui.blit(poseStack, (gui.width / 2) - (ComputerScreen.xSize / 2), yOffset, 0, 0, ComputerScreen.xSize, ComputerScreen.ySize);
-		
 		Font font = Minecraft.getInstance().font;
 		if(be.latestmessage.get(be.programSelected) == null || be.latestmessage.get(be.programSelected).isEmpty())
-			font.draw(poseStack, message, (gui.width - ComputerScreen.xSize) / 2 + 15, (gui.height - ComputerScreen.ySize) / 2 + 45, 4210752);
+			font.draw(poseStack, message, (gui.width - ComputerScreen.xSize) / 2F + 15, (gui.height - ComputerScreen.ySize) / 2F + 45, be.getTheme().getTextColor());
 		else
-			font.draw(poseStack, I18n.get(be.latestmessage.get(be.programSelected)), (gui.width - ComputerScreen.xSize) / 2  + 15, (gui.height - ComputerScreen.ySize) / 2 + 45, 4210752);
+			font.draw(poseStack, I18n.get(be.latestmessage.get(be.programSelected)), (gui.width - ComputerScreen.xSize) / 2F  + 15, (gui.height - ComputerScreen.ySize) / 2F + 45, be.getTheme().getTextColor());
 	}
 	
 	/**
 	 * Represents an unlocalized string and the possible format parameters.
 	 * Is used to represent the value on the buttons, but also the message shown above the buttons.
-	 * See getStringList().
+	 * @see ButtonListProgram#getStringList
 	 */
 	protected static class UnlocalizedString
 	{
@@ -166,4 +162,35 @@ public abstract class ButtonListProgram extends ComputerProgram
 		}
 	}
 	
+	protected class ArrowButton extends ExtendedButton
+	{
+		boolean reverse;
+		ComputerScreen gui;
+		public ArrowButton(boolean reverse, ComputerScreen gui)
+		{
+			super((gui.width-ComputerScreen.xSize)/2 + 140, (gui.height-ComputerScreen.ySize)/2 + (reverse?60:132), 20, 20, Component.empty(), b->onArrowPressed(gui, reverse));
+			this.reverse = reverse;
+			this.gui = gui;
+		}
+		
+		@Override
+		public void renderButton(PoseStack poseStack, int mouseX, int mouseY, float partialTick)
+		{
+			if(active)
+			{
+				RenderSystem.setShader(GameRenderer::getPositionTexShader);
+				RenderSystem.setShaderColor(1, 1, 1, 1);
+				RenderSystem.setShaderTexture(0, gui.be.getTheme().getTexture());
+				blit(poseStack, x, y, 158 + (active ? 0 : 20), reverse ? 0 : 20, 20, 20);
+			} else
+			{
+				// use default minecraft button rendering to draw inactive buttons
+				super.renderButton(poseStack, mouseX, mouseY, partialTick);
+				RenderSystem.setShader(GameRenderer::getPositionTexShader);
+				RenderSystem.setShaderColor(1, 1, 1, 1);
+				RenderSystem.setShaderTexture(0, ComputerScreen.guiMain);
+				blit(poseStack, x, y, reverse?0:20, ComputerScreen.ySize, 20, 20);
+			}
+		}
+	}
 }